doc: Clean up inconsistent documentation comments
We don't (yet) publicly generate documentation based on code
but that doesn't mean the comments have to be useless and
inconsistent.

Changes:
- Removed: @function @name @example
- Using:
  - @class @singleton @method
  - @private @member @alternateClassName
  - @param @return
  - Markdown-compatible syntax in descriptions (e.g. backticks,
    hyperlinks, and things like an empty line being required
    before a list, code indentation etc.).
- @method's name is optional. In most cases the generator
  of choice should be able to infer this from the code.

Though this doesn't mean anything, the above doc tags match a
subset of JSDuck. So if you feel like generating an index
locally, I know it works at least with jsduck. Again, I'm not
recommending we generate this periodically or publish it,
merely using it to enforce some consistency in the comments.

I find this especially useful for getting an insight into
internal code that isn't considered part of the API.